Moved interface to it`s own file, fixed imports
