+++ src/types/RoundPlayer.d.ts
@@ -1,0 +1,8 @@
+import { Player } from "./Player";
+
+export interface RoundPlayer extends Player {
+    cards: string[];
+    cardsScore: number[];
+    hasMadeFinalDecision: boolean;
+    currentStatus: "won" | "lost" | "playing" | "blackjack" | "bust" | "push";
+}
